refactor: PaymentType
