+++ packages/mesh/types/payment.ts
@@ -1,0 +1,6 @@
+export enum PaymentType {
+  SubscriptionStory = 'subscription-story',
+  SubscriptionPublisher = 'subscription-publisher',
+  Sponsor = 'sponsor',
+  Deposit = 'deposit',
+}
